+++ project/persistence/backoffice-jdbc-dao/src/main/java/it/maxmin/dao/jdbc/impl/operation/builder/ResultSetAddressBuilder.java
@@ -17,7 +17,7 @@ public class ResultSetAddressBuilder {
 				address = Address.newInstance().withId(rs.getLong("AddressId"))
 						.withDescription(rs.getString("Description")).withCity(rs.getString("City"))
 						.withPostalCode(rs.getString("PostalCode")).withRegion(rs.getString("Region"))
-						.withVersion(rs.getLong("Version"));
+						.withVersion(rs.getInt("Version"));
 			}
 			return Optional.ofNullable(address);
 		} catch (SQLException e) {
